a few more changes
