+++ .eslintrc
@@ -2,8 +2,7 @@
     "extends": [
         "eslint:recommended",
         "plugin:@typescript-eslint/eslint-recommended",
-        "plugin:@typescript-eslint/recommended",
-        "plugin:prettier/recommended"
+        "plugin:@typescript-eslint/recommended"
     ],
     "plugins": [
         "@typescript-eslint"
@@ -16,17 +15,6 @@
         "ecmaVersion": 2021
     },
     "rules": {
-        "prettier/prettier": [
-            2,
-            {
-                "printWidth": 120,
-                "singleQuote": true,
-                "quoteProps": "as-needed",
-                "trailingComma": "all",
-                "endOfLine": "lf",
-                "arrowParens": "avoid"
-            }
-        ],
         "comma-dangle": ["error", "always-multiline"],
         "comma-spacing": "error",
         "comma-style": "error",
