+++ .npmignore
@@ -1,0 +1,5 @@
+.github/
+images/
+node_modules/
+package-lock.json
+tsconfig.json
