--- .prettierrc
@@ -1,7 +1,0 @@
-{
-    "singleQuote": true,
-    "printWidth": 120,
-    "trailingComma": "all",
-    "endOfLine": "lf",
-    "arrowParens": "avoid"
-}
