+++ src/snakecord.ts
@@ -1,80 +1,106 @@
-const Discord = require('discord.js');
-const WIDTH = 15;
-const HEIGHT = 10;
-const gameBoard = [];
-const apple = { x: 1, y: 1 };
+import { ColorResolvable, Message, MessageEmbed, MessageReaction, User } from 'discord.js';
+import { snakeGameOptions, entityLocation } from '../typings/types';
 
 /**
  * The main class, initializes a new Snake game.
  */
-class SnakeGame {
-    constructor(options = {}) {
+export class SnakeGame {
+    // Board properties
+    boardWidth: number;
+    boardLength: number;
+    gameBoard: string[];
+
+    // Entity properties
+    apple: entityLocation;
+    snake: entityLocation[];
+    snakeLength: number;
+    score: number;
+
+    // Other properties
+    gameEmbed: Message;
+    inGame: boolean;
+    options: snakeGameOptions;
+
+    constructor(options: snakeGameOptions) {
+        this.options = options;
+        this.boardWidth = this.options.width || 15;
+        this.boardLength = this.options.length || 15;
+        this.gameBoard = [];
+        this.apple = { x: 1, y: 1 };
         this.snake = [{ x: 5, y: 5 }];
         this.snakeLength = 1;
         this.score = 0;
         this.gameEmbed = null;
         this.inGame = false;
-        this.options = options;
-        for(let y = 0; y < HEIGHT; y++) {
-            for(let x = 0; x < WIDTH; x++) {
-                gameBoard[y * WIDTH + x] = '🟦';
+
+        for (let y = 0; y < this.boardLength; y++) {
+            for (let x = 0; x < this.boardWidth; x++) {
+                this.gameBoard[y * this.boardWidth + x] = '🟦';
             }
         }
     }
 
     /**
      * Creates a new game board for a Snake game.
-     * @returns A matrix of blue squares and one apple in a random location
      */
-    gameBoardToString() {
+    gameBoardToString(): string {
         let str = '';
-        for(let y = 0; y < HEIGHT; y++) {
-            for(let x = 0; x < WIDTH; x++) {
-                if(x == apple.x && y == apple.y) {
+        for (let y = 0; y < this.boardLength; y++) {
+            for (let x = 0; x < this.boardWidth; x++) {
+                if (x == this.apple.x && y == this.apple.y) {
                     str += '🍎';
                     continue;
                 }
 
                 let flag = true;
-                for(let s = 0; s < this.snake.length; s++) {
-                    if(x == this.snake[s].x && y == this.snake[s].y) {
+                for (let s = 0; s < this.snake.length; s++) {
+                    if (x == this.snake[s].x && y == this.snake[s].y) {
                         str += '🟩';
                         flag = false;
                     }
                 }
 
-                if(flag) {
-                    str += gameBoard[y * WIDTH + x];
+                if (flag) {
+                    str += this.gameBoard[y * this.boardWidth + x];
                 }
             }
+
             str += '\n';
         }
+
         return str;
     }
 
-    isLocInSnake(pos) {
-        return this.snake.find(sPos => sPos.x == pos.x && sPos.y == pos.y);
+    isLocationInSnake(pos: entityLocation): entityLocation {
+        return this.snake.find(snakePos => snakePos.x == pos.x && snakePos.y == pos.y);
     }
 
     /**
      * Moves the apple around the game board.
      */
-    newAppleLoc() {
-        let newApplePos = { x: 0, y: 0 };
-        do {
-            newApplePos = { x: parseInt(Math.random() * WIDTH), y: parseInt(Math.random() * HEIGHT) };
-        } while(this.isLocInSnake(newApplePos));
+    newAppleLocation(): void {
+        let newApplePos = {
+            x: 0,
+            y: 0,
+        };
 
-        apple.x = newApplePos.x;
-        apple.y = newApplePos.y;
+        do {
+            newApplePos = {
+                x: Math.random() * this.boardWidth,
+                y: Math.random() * this.boardLength,
+            };
+        } while (this.isLocationInSnake(newApplePos));
+
+        this.apple.x = newApplePos.x;
+        this.apple.y = newApplePos.y;
     }
 
     /**
      * Creates a new Snake game
-     * @param {Discord.Message} msg
+     * @param {Message} msg - The message instance from which to begin.
      */
-    newGame(msg) {
-        if(this.inGame) {
+    newGame(msg: Message): void {
+        if (this.inGame) {
             return;
         }
 
@@ -82,17 +108,17 @@ class SnakeGame {
         this.score = 0;
         this.snakeLength = 1;
         this.snake = [{ x: 5, y: 5 }];
-        this.newAppleLoc();
-        const embed = new Discord.MessageEmbed()
+        this.newAppleLocation();
+
+        const embed = new MessageEmbed()
             .setColor(this.options.color || 'RANDOM')
             .setTitle(this.options.title || 'Snake Game')
             .setDescription(this.gameBoardToString());
 
-        if(this.options.timestamp) {
+        if (this.options.timestamp) {
             embed.setTimestamp();
         }
 
-        // sending the embed as a message option to prevent edge case errors
         msg.channel.send({ embed }).then(message => {
             this.gameEmbed = message;
             this.gameEmbed.react('⬅️');
@@ -107,91 +133,99 @@ class SnakeGame {
     /**
      * Updates the game board (and the snake if it ate an apple).
      */
-    step() {
-        if(apple.x == this.snake[0].x && apple.y == this.snake[0].y) {
+    step(): void {
+        if (this.apple.x == this.snake[0].x && this.apple.y == this.snake[0].y) {
             this.score += 1;
             this.snakeLength++;
-            this.newAppleLoc();
+            this.newAppleLocation();
         }
 
-        const editEmbed = new Discord.MessageEmbed()
+        const editedEmbed = new MessageEmbed()
             .setColor(this.options.color || 'RANDOM')
             .setTitle(this.options.title || 'Snake Game')
             .setDescription(this.gameBoardToString());
 
-        if(this.options.timestamp) {
-            editEmbed.setTimestamp();
+        if (this.options.timestamp) {
+            editedEmbed.setTimestamp();
         }
 
-        this.gameEmbed.edit(editEmbed);
+        this.gameEmbed.edit(editedEmbed);
         this.waitForReaction();
     }
 
-    gameOver() {
+    gameOver(): void {
         this.inGame = false;
-        const editEmbed = new Discord.MessageEmbed()
+
+        const editedEmbed = new MessageEmbed()
             .setColor(this.options.color || 'RANDOM')
             .setTitle(this.options.gameOverTitle || 'Game Over')
             .setDescription(`SCORE: **${this.score}**`);
 
-        if(this.options.timestamp) {
-            editEmbed.setTimestamp();
+        if (this.options.timestamp) {
+            editedEmbed.setTimestamp();
         }
 
-        this.gameEmbed.edit(editEmbed);
+        this.gameEmbed.edit(editedEmbed);
         this.gameEmbed.reactions.removeAll();
     }
 
     /**
      * The message reaction collector filter.
-     * @param {*} reaction
-     * @param {Discord.User} user
+     * @param {MessageReaction} reaction - The reactions to check for.
+     * @param {User} user - The user who reacted.
      */
-    filter(reaction, user) {
+    filter(reaction: MessageReaction, user: User): boolean {
         return ['⬅️', '⬆️', '⬇️', '➡️'].includes(reaction.emoji.name) && user.id !== this.gameEmbed.author.id;
     }
 
     /**
      * Handles reactions on the game embed.
      */
-    waitForReaction() {
+    waitForReaction(): void {
         this.gameEmbed.awaitReactions((reaction, user) => this.filter(reaction, user), { max: 1, time: 60000, errors: ['time'] }).then(collected => {
             const reaction = collected.first();
-
             const snakeHead = this.snake[0];
-            const nextPos = { x: snakeHead.x, y: snakeHead.y };
-            if(reaction.emoji.name === '⬅️') {
+            const nextPos = {
+                x: snakeHead.x,
+                y: snakeHead.y,
+            };
+
+            if (reaction.emoji.name === '⬅️') {
                 let nextX = snakeHead.x - 1;
-                if(nextX < 0) {
-                    nextX = WIDTH - 1;
+                if (nextX < 0) {
+                    nextX = this.boardWidth - 1;
                 }
+
                 nextPos.x = nextX;
-            } else if(reaction.emoji.name === '⬆️') {
+            } else if (reaction.emoji.name === '⬆️') {
                 let nextY = snakeHead.y - 1;
-                if(nextY < 0) {
-                    nextY = HEIGHT - 1;
+                if (nextY < 0) {
+                    nextY = this.boardLength - 1;
                 }
+
                 nextPos.y = nextY;
-            } else if(reaction.emoji.name === '⬇️') {
+            } else if (reaction.emoji.name === '⬇️') {
                 let nextY = snakeHead.y + 1;
-                if(nextY >= HEIGHT) {
+                if (nextY >= this.boardLength) {
                     nextY = 0;
                 }
+
                 nextPos.y = nextY;
-            } else if(reaction.emoji.name === '➡️') {
+            } else if (reaction.emoji.name === '➡️') {
                 let nextX = snakeHead.x + 1;
-                if(nextX >= WIDTH) {
+                if (nextX >= this.boardWidth) {
                     nextX = 0;
                 }
+
                 nextPos.x = nextX;
             }
 
             reaction.users.remove(reaction.users.cache.filter(user => user.id !== this.gameEmbed.author.id).first().id).then(() => {
-                if(this.isLocInSnake(nextPos)) {
+                if (this.isLocationInSnake(nextPos)) {
                     this.gameOver();
                 } else {
                     this.snake.unshift(nextPos);
-                    if(this.snake.length > this.snakeLength) {
+                    if (this.snake.length > this.snakeLength) {
                         this.snake.pop();
                     }
 
@@ -205,18 +239,18 @@ class SnakeGame {
 
     /**
      * Sets the custom (or default) title of the game embed.
-     * @param {String} title
+     * @param {String} title - The title to set.
      */
-    setTitle(title) {
+    setTitle(title: string): this {
         this.options.title = title || 'Snake: The Game';
         return this;
     }
 
     /**
      * Sets the custom (or default) color of the game embed.
-     * @param {*} color
+     * @param {ColorResolvable} color - The hex code for the embed color.
      */
-    setColor(color) {
+    setColor(color: ColorResolvable): this {
         this.options.color = color || 'RANDOM';
         return this;
     }
@@ -224,10 +258,8 @@ class SnakeGame {
     /**
      * Sets the timestamp of the game embed if requested.
      */
-    setTimestamp() {
+    setTimestamp(): this {
         this.options.timestamp = true;
         return this;
     }
 }
-
-module.exports = SnakeGame;
