+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2019",
+        "declaration": true,
+        "outDir": "./out"
+    },
+    "include": [
+        "src/**/*"
+, "typings/types.d.ts"    ]
+}
